fixed the daily exopense reminder feature
